fixed deg/rad problems and refactored rendering

+++ gym-ballhoop/gym_ballhoop/envs/update.py
@@ -36,7 +36,7 @@ def update_outer(state, t_i):
     Dphi_out = (Dth_out - Dpsi_out) * (params.Ro / params.Rb)
     mode_out = 1
 
-    return np.asarray([th_out, Dth_out, psi_out, Dpsi_out, r_out, Dr_out, phi_out, Dphi_out, mode_out])
+    return np.asarray([th_out, Dth_out, psi_out, Dpsi_out, r_out, Dr_out, phi_out, Dphi_out, mode_out], dtype=np.float64)
 
 def update_free(state, t_i):
     '''
@@ -74,7 +74,7 @@ def update_free(state, t_i):
     Dpsi_out = Dpsi + params.h * Dpsi_dot
     mode_out = 2
 
-    return np.asarray([th_out, Dth_out, psi_out, Dpsi_out, r_out, Dr_out, phi_out, Dphi_out, mode_out])
+    return np.asarray([th_out, Dth_out, psi_out, Dpsi_out, r_out, Dr_out, phi_out, Dphi_out, mode_out], dtype=np.float64)
 
 def update_inner(state, t_i):
     '''
@@ -108,7 +108,7 @@ def update_inner(state, t_i):
     Dphi_out = -(Dth_out - Dpsi_out) * (params.Rui / params.Rb)
     mode_out = 3
 
-    return np.asarray([th_out, Dth_out, psi_out, Dpsi_out, r_out, Dr_out, phi_out, Dphi_out, mode_out])
+    return np.asarray([th_out, Dth_out, psi_out, Dpsi_out, r_out, Dr_out, phi_out, Dphi_out, mode_out], dtype=np.float64)
 
 def update_all(state, t_i, double_hoop=False,debug=False):
     '''
